Moved the mock todo dao storing
