+++ test/di/mock/mock_todo_dao_storing.dart
@@ -1,0 +1,27 @@
+import 'package:flutter_template/database/flutter_template_database.dart';
+import 'package:flutter_template/database/todo/todo_dao_storing.dart';
+import 'package:flutter_template/model/webservice/todo/todo.dart';
+import 'package:mockito/mockito.dart';
+
+class MockTodoDaoStoring extends Mock implements TodoDaoStoring {
+  @override
+  Future<void> createTodoWithValue(Todo? todo) async {
+    noSuchMethod(Invocation.method(#createTodoWithValue, [todo]));
+  }
+
+  @override
+  Stream<List<DbTodo>> getAllTodosStream() {
+    // ignore: avoid_as
+    return noSuchMethod(Invocation.method(#getAllTodosStream, []), returnValue: Stream.value(<DbTodo>[])) as Stream<List<DbTodo>>;
+  }
+
+  @override
+  Future<void> createTodo(String todo) async {
+    noSuchMethod(Invocation.method(#createTodo, [todo]));
+  }
+
+  @override
+  Future<void> updateTodo({required int id, required bool completed}) async {
+    noSuchMethod(Invocation.method(#updateTodo, [id, completed]));
+  }
+}
